requirement.txt
flask==2.3.3
langchain==0.1.0
langchain-openai==0.0.5
langchain-core==0.1.0
pydantic==1.10.13
python-dotenv==1.0.0